--- a/mujoco/mjx/_src/constraint.py
+++ b/mujoco/mjx/_src/constraint.py
@@ -26,8 +26,8 @@ def _update_efc_row(
   pos_aref: wp.float32,
   pos_imp: wp.float32,
   invweight: wp.float32,
-  solref: wp.array(ndim=1, dtype=wp.float32),
-  solimp: wp.array(ndim=1, dtype=wp.float32),
+  solref: wp.vec2f,
+  solimp: types.vec5,
   margin: wp.float32,
   refsafe: bool,
   Jqvel: float,
@@ -145,22 +145,26 @@ def _efc_contact_pyramidal(
   d: types.Data,
   refsafe: bool,
 ):
-  worldid, conid, dimid = wp.tid()
+  conid, dimid = wp.tid()
 
-  if d.contact.dim[worldid, conid] != 3:
+  if conid >= d.ncon_total[0]:
+    return
+  
+  if d.contact.dim[conid] != 3:
     return
 
-  pos = d.contact.dist[worldid, conid] - d.contact.includemargin[worldid, conid]
+  pos = d.contact.dist[conid] - d.contact.includemargin[conid]
   active = pos < 0
 
   if active:
     efcid = wp.atomic_add(d.nefc_total, 0, 1)
+    worldid = d.contact.worldid[conid]
     d.efc_worldid[efcid] = worldid
 
-    body1 = m.geom_bodyid[d.contact.geom[worldid, conid, 0]]
-    body2 = m.geom_bodyid[d.contact.geom[worldid, conid, 1]]
+    body1 = m.geom_bodyid[d.contact.geom[conid][0]]
+    body2 = m.geom_bodyid[d.contact.geom[conid][1]]
 
-    fri0 = d.contact.friction[worldid, conid, 0]
+    fri0 = d.contact.friction[conid][0]
 
     # pyramidal has common invweight across all edges
     invweight = m.body_invweight0[body1, 0] + m.body_invweight0[body2, 0]
@@ -174,16 +178,16 @@ def _efc_contact_pyramidal(
       diff_0 = float(0.0)
       diff_i = float(0.0)
       for xyz in range(3):
-        con_pos = d.contact.pos[worldid, conid]
+        con_pos = d.contact.pos[conid]
         jac1p = _jac(m, d, con_pos, xyz, body1, i, worldid)
         jac2p = _jac(m, d, con_pos, xyz, body2, i, worldid)
         jac_dif = jac2p - jac1p
-        diff_0 += d.contact.frame[worldid, conid][0, xyz] * jac_dif
-        diff_i += d.contact.frame[worldid, conid][dimid2, xyz] * jac_dif
+        diff_0 += d.contact.frame[conid][0, xyz] * jac_dif
+        diff_i += d.contact.frame[conid][dimid2, xyz] * jac_dif
       if dimid % 2 == 0:
-        J = diff_0 + diff_i * d.contact.friction[worldid, conid, dimid2 - 1]
+        J = diff_0 + diff_i * d.contact.friction[conid][dimid2 - 1]
       else:
-        J = diff_0 - diff_i * d.contact.friction[worldid, conid, dimid2 - 1]
+        J = diff_0 - diff_i * d.contact.friction[conid][dimid2 - 1]
 
       d.efc_J[efcid, i] = J
       Jqvel += J * d.qvel[worldid, i]
@@ -196,9 +200,9 @@ def _efc_contact_pyramidal(
       pos,
       pos,
       invweight,
-      d.contact.solref[worldid, conid],
-      d.contact.solimp[worldid, conid],
-      d.contact.includemargin[worldid, conid],
+      d.contact.solref[conid],
+      d.contact.solimp[conid],
+      d.contact.includemargin[conid],
       refsafe,
       Jqvel,
     )
@@ -225,6 +229,6 @@ def make_constraint(m: types.Model, d: types.Data):
     if m.opt.cone == types.ConeType.PYRAMIDAL.value:
       wp.launch(
         _efc_contact_pyramidal,
-        dim=(d.nworld, d.ncon, 4),
+        dim=(d.nconmax, 4),
         inputs=[m, d, refsafe],
       )