`test_util.benchmark` was temporarily modified to load a keyframe, call `mj_forward`, and then create `Data` using `put_data` in order to generate these benchmarking results.

```
d = mujoco.MjData(m)
mujoco.mj_resetDataKeyframe(m, d, 0)
mujoco.mj_forward(m, d)
mx = io.put_model(m)
dx = io.put_data(m, d)
```

keyframe `0` for the humanoid results in `nefc=32` (comprising four pyramidal friction contacts with condim=3)

```
mjx-testspeed --function=make_constraint --mjcf=humanoid/humanoid.xml --batch_size=8192
```

this pr:
```
Summary for 8192 parallel rollouts

 Total JIT time: 1.47 s
 Total simulation time: 0.06 s
 Total steps per second: 136,651,564
 Total realtime factor: 683,257.82 x
 Total time per step: 0.01 µs
```

main:
```
Summary for 8192 parallel rollouts

 Total JIT time: 1.51 s
 Total simulation time: 0.06 s
 Total steps per second: 130,614,030
 Total realtime factor: 653,070.15 x
 Total time per step: 0.01 µs
```