@erikfrey 2025-02-26
@thowell generally looks great but can you update the PR description with before/after `mjx-benchmark` results for `make_constraint` ?