@thowell 2025-02-27
@erikfrey the pr description is updated with initial benchmarking results